backend/services/payout-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic>=2.12.0
pydantic-settings>=2.11.0
PyJWT==2.10.1
psycopg[binary]==3.2.10
kombu==5.5.4
prometheus-client==0.22.1
python-json-logger==3.3.0
httpx==0.28.1
pytest==8.4.2
pytest-cov==7.0.0
locust==2.41.6
